+++ src/utils/FormElementMixin.js
@@ -37,8 +37,16 @@ export default {
          */
         statusType() {
             if (!this.parentField) return
-
-            return this.parentField.newType
+            if (!this.parentField.newType) return
+            if (typeof this.parentField.newType === 'string') {
+                return this.parentField.newType
+            } else {
+                for (let key in this.parentField.newType) {
+                    if (this.parentField.newType[key]) {
+                        return key
+                    }
+                }
+            }
         },
 
         /**